+++ FreeCamClass.cs
@@ -15,6 +15,7 @@ namespace SimpleFreeCam
 
         private static PlayerControllerB playerController = null;
         internal static bool isInVehicle = false;
+        internal static bool shouldChangeFOV = false;
 
         private static Transform mainCameraTransform;
         private static Transform freeCameraTransform;
@@ -25,7 +26,13 @@ namespace SimpleFreeCam
         internal static bool isInFreeCam = false;
 
         internal static float freecamSpeed = 3f;
-        internal static bool hasBootedAlready = false;
+        private static bool hasBootedAlready = false;
+
+        private static float PlayerSavedCameraUp = 0f;
+        public static float FreeCamSavedCameraUp = 0f;
+        private static float freecamFOV = 60f;
+
+        private static readonly FieldInfo cameraUpField = typeof(PlayerControllerB).GetField("cameraUp", BindingFlags.NonPublic | BindingFlags.Instance);
 
         public static IEnumerator SetupFreecam()
         {
@@ -62,6 +69,16 @@ namespace SimpleFreeCam
             }
         }
 
+        public static void ToggleFOVSwitch_performed(bool isHeldDown)
+        {
+            if (playerController == null || !isInFreeCam)
+            {
+                return;
+            }
+
+            shouldChangeFOV = isHeldDown;
+        }
+
         public static void ToggleFreecamMovement_performed(InputAction.CallbackContext obj)
         {
             if (!obj.performed)
@@ -79,6 +96,43 @@ namespace SimpleFreeCam
             DisableFreecamOnLock();
         }
 
+        public static void ResetFOV_performance(InputAction.CallbackContext obj)
+        {
+            if (!obj.performed)
+            {
+                return;
+            }
+
+            if (playerController == null || !isInFreeCam)
+            {
+                return;
+            }
+
+            freecamFOV = FreeCamFOVInfo.DefaultFOV;
+            FreeCam.fieldOfView = FreeCamFOVInfo.DefaultFOV;
+        }
+
+        public static void ResetFreeCam()
+        {
+            hasBootedAlready = false;
+            FreeCamSavedCameraUp = 0f;
+        }
+
+        public static void TeleportFreeCamToPlayer_performance(InputAction.CallbackContext obj)
+        {
+            if (!obj.performed)
+            {
+                return;
+            }
+
+            if (playerController == null || !isInFreeCam)
+            {
+                return;
+            }
+
+            TeleportFreeCamToPlayer();
+        }
+
         public static void SetPlayerController(PlayerControllerB controller)
         {
             playerController = controller;
@@ -89,20 +143,87 @@ namespace SimpleFreeCam
             isInVehicle = inVehicle;
         }
 
+        public static void SetFreecamLocation(Vector3 position)
+        {
+            if (isInFreeCam && !lockFreeCam)
+            {
+                freeCameraTransform.position = position;
+                StartOfRound.Instance.freeCinematicCameraTurnCompass.position = position;
+            }
+        }
+
+        public static void SetFreecamFOV(float fov)
+        {
+            freecamFOV = Mathf.Clamp(fov, FreeCamFOVInfo.MinFOV, FreeCamFOVInfo.MaxFOV);
+            if (isInFreeCam) FreeCam.fieldOfView = freecamFOV;
+        }
+
+        public static float GetFreecamFOV() => freecamFOV;
+
+        public static void TeleportFreeCamToPlayer()
+        {
+            TeleportFreeCamToPlayer(-0.5f);
+        }
+
+        private static void ResetFreeCamRotation()
+        {
+            Quaternion newRotation;
+            if (playerController != null)
+            {
+                Vector3 forward = playerController.gameplayCamera.transform.forward;
+                forward.y = 0;
+                forward.Normalize();
+
+                newRotation = Quaternion.LookRotation(forward, Vector3.up);
+            } else
+            {
+                newRotation = default;
+            }
+
+            freeCameraTransform.rotation = newRotation;
+            StartOfRound.Instance.freeCinematicCameraTurnCompass.rotation = newRotation;
+        }
+
+        public static void TeleportFreeCamToPlayer(float offset)
+        {
+            Vector3 playerCamPosition = playerController.gameplayCamera.transform.position;
+            playerCamPosition += playerController.gameplayCamera.transform.forward * offset;
+            SetFreecamLocation(playerCamPosition);
+        }
+
         private static void EnableFreecam()
         {
             //alays start with freecamMovement true
             lockFreeCam = false;
             isInFreeCam = true;
+            shouldChangeFOV = false;
 
-            //reset each time on enable
-            if (SimpleFreeCamPatchBase.instance.FreeCamConfigEntry.Value || !hasBootedAlready)
+            FreeCam.cullingMask = playerController.gameplayCamera.cullingMask;
+
+            if (!hasBootedAlready)
             {
                 hasBootedAlready = true;
                 freecamSpeed = SimpleFreeCamPatchBase.instance.FreeCamConfigEntryFloat.Value;
+
+                //teleport freecam to player on first "boot"
+                TeleportFreeCamToPlayer(-1);
+                //reset rotation
+                ResetFreeCamRotation();
             }
 
+            //reset each time on enable
+            if (SimpleFreeCamPatchBase.instance.FreeCamConfigEntry.Value)
+            {
+                freecamSpeed = SimpleFreeCamPatchBase.instance.FreeCamConfigEntryFloat.Value;
+            }
+
+            SwitchCameraUpField(true);
+
+            FreeCam.fieldOfView = freecamFOV;
+
             playerController.isFreeCamera = true;
+
+            StartOfRound.Instance.freeCinematicCameraTurnCompass.rotation = freeCameraTransform.rotation;
             StartOfRound.Instance.SwitchCamera(FreeCam);
 
             HUDObject.SetActive(false);
@@ -113,10 +234,36 @@ namespace SimpleFreeCam
             HUDManager.Instance.HideHUD(true);
         }
 
+        private static void SwitchCameraUpField(bool switchToFreeCam)
+        {
+            //should the freecam pitch be used
+            if (switchToFreeCam)
+            {
+                //save player character pitch
+                PlayerSavedCameraUp = (float)cameraUpField.GetValue(playerController);
+
+                //set camera pitch to freecam pitch
+                cameraUpField.SetValue(playerController, FreeCamSavedCameraUp);
+            } else
+            {
+                //save freecam pitch
+                FreeCamSavedCameraUp = (float)cameraUpField.GetValue(playerController);
+
+                //set to player character pitch
+                cameraUpField.SetValue(playerController, PlayerSavedCameraUp);
+            }
+        }
+
         public static void DisableFreecam()
         {
+            if (!lockFreeCam)
+            {
+                SwitchCameraUpField(false);
+            }
+
             lockFreeCam = false;
             isInFreeCam = false;
+
 
             playerController.isFreeCamera = false;
             StartOfRound.Instance.freeCinematicCamera.enabled = false;
@@ -133,14 +280,8 @@ namespace SimpleFreeCam
         public static void DisableFreecamOnLock()
         {
             isInFreeCam = true;
-            if (lockFreeCam)
-            {
-                playerController.isFreeCamera = false;
-            }
-            else
-            {
-                playerController.isFreeCamera = true;
-            }
+            SwitchCameraUpField(!lockFreeCam);
+            playerController.isFreeCamera = !lockFreeCam;
         }
 
         internal static void AdjustSpeed(float scrollDelta)
@@ -151,5 +292,16 @@ namespace SimpleFreeCam
         }
 
         internal static float GetSpeed() => freecamSpeed;
+
+        internal static void AdjustFOV(float delta)
+        {
+            if (!isInFreeCam || lockFreeCam) return;
+            SetFreecamFOV(freecamFOV + delta);
+        }
+
+        internal static float GetFOVSensitivityMultiplier()
+        {
+            return freecamFOV / FreeCamFOVInfo.DefaultFOV;
+        }
     }
 }