+++ FreeCamFOVInfo.cs
@@ -1,0 +1,10 @@
+﻿namespace SimpleFreeCam
+{
+    internal static class FreeCamFOVInfo
+    {
+        internal const float MinFOV = 5f;
+        internal const float MaxFOV = 160f;
+        internal const float ScrollStep = 1f;
+        internal const float DefaultFOV = 60f;
+    }
+}