+++ ModInfo.cs
@@ -4,6 +4,6 @@
     {
         internal const string modGUID = "PixelIndieDev_SimpleFreeCam";
         internal const string modName = "Simple Freecam";
-        internal const string modVersion = "1.0.0.0";
+        internal const string modVersion = "1.1.1.0";
     }
 }