+++ CHANGELOG.md
@@ -1,2 +1,14 @@
+## v1.1.1
+- Fixed readme formatting
+
+## v1.1.0
+- Added an option to adjust the freecam field of view (FOV)
+- Added a keybind to teleport the freecam to the player
+- Added a keybind to reset the freecam FOV
+- The Freecam now spawns behind the player on first use
+- The Freecam look input is no longer smoothed, making camera look movement match the normal player camera
+- Fixed an issue where some items were not visible while in freecam
+- Fixed an issue where the freecam pitch would sync with the player character's pitch when toggling the freecam
+
 ## v1.0.0
 - Initial release